refactor the App object to a normal object instead of inline-defined object
